in progress for spec02_02

+++ ObjectsRelated/Pickupables/Subs/_CusotmDefinedSingleUse/Pickupable_CustomDefinedSingleUse.gd
@@ -1,5 +1,8 @@
 extends "res://ObjectsRelated/Pickupables/BasePickupables.gd"
 
+
+var is_destroy_self_on_player_entered : bool = true
+
 signal player_entered_self__custom_defined()
 signal restored_from_destroyed_from_rewind()
 
@@ -8,7 +11,8 @@ func _on_player_entered_self(arg_player):
 	
 	emit_signal("player_entered_self__custom_defined")
 	
-	_destroy_self__on_consume_by_player()
+	if is_destroy_self_on_player_entered:
+		_destroy_self__on_consume_by_player()
 
 
 func restore_from_destroyed_from_rewind():
